+++ src/parts/component.js
@@ -7,8 +7,9 @@ export function makeComponent(node, option={}) {
 
   this.require('$context');
 
+  let deepChecking = this.config.deepCheckingProps;
   let reference = null;
-  let propsFn = [], propsSetter = [], $class = [], staticProps = true, deepChecking = false;
+  let propsFn = [], propsSetter = [], $class = [], staticProps = true;
   let slotBlocks = [];
   let anchorBlocks = [];
 
@@ -159,6 +160,12 @@ export function makeComponent(node, option={}) {
       let inner, outer;
       if(name[0] == ':') inner = name.substring(1);
       else inner = name.substring(5);
+      let mods = inner.split('|');
+      inner = mods.shift();
+      mods.forEach(mod => {
+        if (mod == 'deep') deepChecking = true;
+        else throw new Error('Wrong modifier: ' + mod);
+      });
       if(value) outer = unwrapExp(value);
       else outer = inner;
       assert(isSimpleName(inner), `Wrong property: '${inner}'`);